requirements.txt
fastapi==0.58.1
uvicorn==0.11.5
aiokafka==0.6.0
confluent-kafka==1.9.2
httptools==0.1.2
iniconfig==1.1.1
kafka-python==2.0.2
packaging==21.3
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.3
requests==2.28.1
starlette==0.13.4
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.11.5
uvloop==0.17.0
websockets==8.1